+++ src/workspace.h
@@ -2,6 +2,7 @@
 #define WORKSPACE_H
 
 #define SUPPORTED_MACOS_VERSION_LIST    \
+    SUPPORT_MACOS_VERSION(goldengate, 27) \
     SUPPORT_MACOS_VERSION(tahoe,    26) \
     SUPPORT_MACOS_VERSION(sequoia,  15) \
     SUPPORT_MACOS_VERSION(sonoma,   14) \